Reorganize database related files
